[argparse] better argparsing seeking .also(UPDATE): utils header inclusion in sources .state: in the middle of refactoring promises for `sdread` .NOTE: these sources will probably be a part of `cproj`        as it will be really annoying to have to rewrite        this from scratch again

+++ utils/err/err.h
@@ -31,7 +31,7 @@ struct error_info {
 typedef struct file_info FInfo;
 typedef struct ferror_info FEInfo;
 
-typedef long (*function) (void);
+typedef int (*function) (void);
 
 extern function const header;
 extern function const body;
